Fixes
There should no longer be a duplicate key error. I think it had something to do with the storage process for victors storing them "with" an _id key already and trying to give them a new one.

There is now a delay before death, and players can't do anything during that time. It is there to allow for their character to enter the podium.